+++ SchemaRefinery/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'UMMI'
 __email__ = 'imm-bioinfo@medicina.ulisboa.pt'
-__version__ = '0.4.0'
+__version__ = '0.5.0'